kpekpe-backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
psycopg2-binary
python-dotenv
gunicorn
uvicorn
